جاري العمل على المرحلة الثانية من تطوير اللغة
تم تمرير &input_ في المعرب والمحلل

جاري العمل على ملفات الخطأ
